requirements.txt
pandas~=2.2.3
requests~=2.32.0
plotly~=5.18.0
python-dateutil~=2.9.0.post0
numpy~=2.2.5
scipy~=1.15.2